Move the ElementFactory and add default values to elements
