+++ src/draw/index.js
@@ -1,7 +1,6 @@
 import Styles from 'diagram-js/lib/draw/Styles';
 import { PluginRegisterBehavior } from './behaviors/PluginRegisterBehavior';
 import { SetRootElementBehavior } from './behaviors/SetRootElementBehavior';
-import { DefaultFactory } from './DefaultFactory';
 import { Renderer } from './Renderer';
 
 
@@ -17,5 +16,4 @@ export default {
     ],
     styles: [ 'type', Styles ],
     renderer: [ 'type', Renderer ],
-    defaultFactory: [ 'type', DefaultFactory ],
 };
